Rewrote 5.3.1.  Discovery of Registrar by Registrar-Agent because it was incorrect and incomplete:
> The discovery of the domain registrar may be done as specified in {{RFC8995}} section 4 with the
> deviation that it is done between the registrar-agent and the domain registrar.
> Alternatively, the registrar-agent may be configured with the address of the domain registrar and the certificate of the domain registrar.

Issues: we need not only discovery, but also selection. Also, this was not in section 4. Lastly, discovery of registrar does not change the need to configure/learn the certificate of the registrar in another way because discovery is not trustworthy (except over ACP, but thats just an instance of how you can pre-learn a trust anchor for the domain registrar(s)).

Fix got somewhat longer, but i tried to ensure all the interop important aspects are covered. Alas, thats a range of details.

Also fixed/added required references and IANA allocations and made the document an update to RFC8995 as indicated in the new text.

This is 50% of the github issue, the other one is 5.3.2 next.